Use int32_t for progress report

+++ src/OpenLoco/src/Objects/ObjectIndex.cpp
@@ -293,7 +293,7 @@ namespace OpenLoco::ObjectManager
         _installedObjectCount = 0;
         // Create new index by iterating all DAT files and processing
         IndexHeader header{};
-        uint8_t progress = 0;      // Progress is used for the ProgressBar Ui element
+        int32_t progress = 0;      // Progress is used for the ProgressBar Ui element
         size_t usedBufferSize = 0; // Keep track of used space to allow for growth and for final sizing
         const auto objectPath = Environment::getPathNoWarning(Environment::PathId::objects);
         for (const auto& file : fs::directory_iterator(objectPath, fs::directory_options::skip_permission_denied))
